[Rename] weather 파일 service 폴더로 이동
