--- services/testBoards.json
@@ -1,7 +1,0 @@
-{
-  "title": "test2",
-  "content": "test3",
-  "name": "annoy",
-  "password": "KK1@@314",
-  "weather": "화창함"
-}
